requirements.txt
fastapi
uvicorn[standard]
aiogram==3.13.1
python-dotenv